requirements.txt
Flask==2.2.5
mysql-connector-python==8.0.33
Flask-SocketIO==5.3.6
python-socketio==5.8.0
requests==2.31.0
python-dotenv==1.0.0
